requirement.txt
aliyun-python-sdk-core==2.8.2
aliyun-python-sdk-sts==3.0.0
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.4.16
chardet==3.0.4
crcmod==1.7
cryptography==2.4.1
defusedxml==0.5.0
Django==1.10
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.4.4
djangorestframework-xml==1.3.0
et-xmlfile==1.0.1
idna==2.6
jdcal==1.4
kombu==4.1.0
MySQL-python==1.2.5
numpy==1.15.1
openpyxl==2.5.6
oss2==2.4.0
pandas==0.23.4
paramiko==2.4.2
Pillow==5.2.0
pyasn1==0.4.4
pycparser==2.19
pycryptodome==3.6.1
PyNaCl==1.3.0
python-dateutil==2.7.3
python-redis==0.2.1
pytz==2018.4
redis==2.10.6
requests==2.18.4
scipy==1.1.0
urllib3==1.22
uWSGI==2.0.17
vine==1.1.4
xlrd==1.1.0
xlwt==1.3.0
